Refactor attachment converter logic, support HEICS for #346
