+++ imessage-exporter/src/app/compatibility/mod.rs
@@ -1,3 +1,3 @@
 pub mod attachment_manager;
-pub mod convert;
+pub mod converters;
 pub mod models;
