+++ imessage-exporter/src/app/mod.rs
@@ -1,6 +1,6 @@
 pub mod error;
 pub mod export_type;
-pub mod converters;
+pub mod compatibility;
 pub mod options;
 pub mod progress;
 pub mod runtime;
